requirements.txt
asgiref==3.3.4
click==7.1.2
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
fastapi==0.63.0
fastapi-login==1.6.0
h11==0.12.0
httptools==0.1.1
importlib-metadata==4.0.1
Markdown==3.3.4
passlib==1.7.4
pydantic==1.8.1
PyJWT==2.0.1
python-dotenv==0.17.0
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
zipp==3.4.1
